Refactor topics/numerals into various numbers.py files in mobject/, animation/, and continual_animation/
